Final sync of jparse before code freeze
.. or at least it's hopefully the last sync :-)

Sync jparse from jparse repo (https://github.com/xexyl/jparse/). This
cleans up some code, removes code that's unnecessary and syncs the versions of
all tools to the same: 1.2.0 2024-10-09. A new version string was added as
well, JPARSE_UTF8_VERSION.

A bug in a script was also fixed. As the use of this script in this repo
(in soup/) already uses the option to set the path of the tool and the
correct path is set already this should not be a problem.

Run make release to test this under macOS.