Update Makefile.example to make more like winners
To help the judges in integrating winning entries to the winner repo,
the Makefile.example file now shows how to include alt code.

It also points out that for variables one should put a space after the
'=' unless it is empty.

Disable -g3. I have a lot of experience with -g and it never works well
if optimisation is enabled (and it's -O3). Also, in macOS it makes the
annoying .dSYM directory.

It points out that for the try rule if the program does not need any arg
you do not have to specify any.

Updated .gitignore to have '.*.sw[a-zA-Z0-9]', which was discovered as
missing when I had Makefile.example open when committing.