Bad typo fix in prep.sh, improve test workflow
The typo was that the variable LOGFILE had no G.

The workflow now runs make bug_report. This does run make test again but
it seems like a good thing to run as it can help uncover more problems
that might not be noticed otherwise. The bug_report.sh script is slow
but this seems like a worthy thing to do.