Sync jparse from jparse repo
This includes some important bug fixes, namely that some codepoints that
were rejected are no longer rejected as later Unicode standards allow
them. This means that the JSONTestSuite
(https://github.com/nst/JSONTestSuite) files now completely pass jparse.

The CODE_OF_CONDUCT.md was removed from jparse/ as it referred to the IOCCC
judges which is not correct. Until a better way of doing this is derived, this
file no longer exists. To be clear the mkiocccentry CODE_OF_CONDUCT.md still
exists!