Make important updates to remarks.example.md
Links were incorrect.

Link to FAQs about certain things.

Stress to pay attention to the rules and guidelines, and especial
attention to rule 17 (these are all linked to).

The links refer to the temp-test-ioccc website so once the Great Fork
Merge happens these links will also have to be updated to refer to the
official IOCCC website.