Sync jparse/ with some important updates
Unused code (that we thought might be needed but seem to not) was
removed.

Extra sanity checks were added to a function.

A file was renamed and this was done here as well to match that.

The CHANGES.md file was corrected to have a missing release so that
there's no confusion on the jump in release numbers.