Various fixes - sync jparse, others
Fixes in jparse/ from the jparse repo (https://github.com/xexyl/jparse/)
that were not applied or even done until a short bit ago.

Typo fix in mkiocccentry.c.

Also roll back a commit in chkentry.c - a work in progress for after IOCCC28
that was, in a rush to leave the house but still get important fixes in, was
accidentally committed yesterday.