Sync jparse/ from jparse repo
This adds a new utility function that will be used in required updates
for chkentry(1) post IOCCC28 and updates the util test code (which tests
the new function).

A minor change in the jstrencode.1 man page was also done though wording
might be needing in fixing (that can come another time however as other
changes might be required).