Fix critical bug in jparse/test_jparse/Makefile
A note is that due to an accidental commit to master branch (in my fork)
it is proving a problem to update the jparse/ directory without the
changes after the code freeze *here* from the jparse repo. It is hoped
that since the state of jparse/, again *here*, is from right before the
code freeze *here*, that there won't actually be changes in the
mkiocccentry repo so that the only file that changes *there* is the
jparse/test_jparse/Makefile. If this does not happen then Landon can add
just the change in the test_jparse/Makefile.

This critical fix is that in Red Hat systems the test_jparse/util_test
cannot compile and without this Red Hat systems will not compile the
jparse library due to the compilation failure.