Revert "Fix critical bug in jparse/test_jparse/Makefile"
This reverts commit f54ea33273bd4ebfc60e007d6bfb387e32b0a66d.

Explanation: this commit was done in error and since it has caused a
problem on my end, being post the code freeze (though due to new things
discovered the code will be unfrozen for a bit) and not merged, I cannot
use my fork to add changes without this. Also, it seems like the
compilation error no longer is a problem (although this has not been
tested in strictly jparse, it does not hurt things and it'll eventually
be synced here) and finally because there was a bug introduced (in the
jstrdecode -h option, though this has been fixed already in jparse).