Fix issues #1036, #1035 and #1033
The issues in respective order:

- Typo fix in mkiocccentry.c ('You title must ...' -> 'Your title must
...').
- mkiocccentry UUID type 4 was too strict
- Renamed makefile.local -> Makefile.local. This does NOT address issue #1034!

Updated version of mkiocccentry due to the above changes. For the second
one new macros are in soup/limit_ioccc.h so the soup version has been
updated as well.

Due to the mkiocccentry version change many JSON files were updated so
this is a large commit.

Something that should be pointed out is that the dyn_array and dbg repos
should be updated to match the Makefile changes. This can be done later.
I am too tired to do that now but what this means is that if one were to
sync dbg or dyn_array to here it would roll it back.

Another thing to point out is that another couple commits will likely be
included as a mistake was made on my end and unfortunately that commit
was pushed to my fork. This means I had to do git revert on it.