Resolve issue #1037
The submit server details are now ready!

Also change the temp-test-ioccc URLs to the main IOCCC website, for once the
Great Fork Merge is performed and the IOCCC website is updated. Although
the links should be okay it would still be wise to, once the merge is
done, test them all. They are in mkiocccentry.h.

Updated mkiocccentry version to: "1.1.0 2024-12-28".