Fix jparse rule rebuild_jparse_err_files
Fix typo in jparse rebuild_jparse_err_files rule that was causing a
display problem. Also here the -v flag cannot depend on the VERBOSITY level as
the script was designed specifically for -v 0! This has been fixed.

The rebuild_jparse_err_files rule HERE must also use the new rule in
test_ioccc/Makefile as we have additional error location files to test.
Although only Landon or I (Cody) will ever need to do this and it's also
extremely unlikely it'll have to happen again, it is still a needed
update.

This commit should fix the workflow error in the previous commit
33951ba1e3b7765eae3028034f491b40a9bed74a.