Move party.json{,.err} to jparse repo
This had to be done due to the improvement with yyerror() and how the
error messages have to be exact. It might be possible to fix the issue
but it is not worth it as long as jparse here tests location errors and
it now does. Plus party.json belongs in jparse anyway.

What this commit means is that the files were moved from
test_ioccc/test_JSON/general.json/bad_loc to
jparse/test_jparse/test_JSON/bad_loc.

The top level Makefile and the test_ioccc/Makefile both had to be
updated but that's only because we no longer need the error location
files in test_ioccc/test_JSON/general.json/bad_loc so the relevant rule
no longer calls the no longer existing rule rebuild_jparse_err_files.

Updated test_ioccc/ioccc_test.sh to use option -L in jparse_test.sh
as we no longer want to test error location messages. This is the only
functional difference in this commit.

Updated ioccc_test.sh version to: "1.0.3 2024-12-31".