Disable install -v by default in install rule
To be more portable we cannot have install -v the default as not all
install(1) implementations support it.

If one needs this functionality they can do:

    make INSTALL_V=-v install

The jparse repo was synced with this change. The dbg and dyn_array repos
have not been updated yet but as the Makefile does not run make install
in either of those directories (or for that matter jparse) it's okay -
for the moment at least.