Minor fixes
Update regexp error message in mkiocccentry. (This was done in jparse too and
that was synced from the jparse repo).

Fix check in mkiocccentry (<= 0, not < 0).

Minor bug fix in txzchk where errno was set to 0 after the function call.

Updated txzchk and mkiocccentry version which required updating JSON files too.

Slight update to GitHub issue template files to match the other one (and should
look nicer in the subject line).