Improve and rename test_extra_file()
Renamed test_extra_file() to better account for what it is:
test_extra_filename(). Improved the function to use two char *[]s so that we
don't have to have checks on each file manually in each tool that uses it,
though at present only chkentry(1) uses it (it is not clear if any other tool
can use it in the future but this update allows for more easily maintaining the
list of filenames that must be present and that must not be present).