Bug fix txzchk for required files
Bug fix txzchk to check for required files in top level directory only (with
the exception that dot files are not allowed in any subdirectory and the only
dot files allowed in the top level directory are .info.json and .auth.json).

Added new test files to txzchk test suite, both good and bad (with associated
.err files).

This fix involved a bug fix and enhancement to the jparse library (not the JSON
parser but the util.c file) which involved syncing the jparse repo again.

Updated TXZCHK_VERSION to "1.1.4 2025-01-17".