Bug fix bug_report.sh to check Makefile.local
Bug fix bug_report.sh to check for Makefile.local, not makefile.local.
This same change was done in other files that still referred to it with the
exception of this file and .gitignore. It was done in dbg/ and dyn_array/ but
the change needs to be done in the repos too.