Fix inspect_Makefile() in mkiocccentry for rule 17
If someone uses a built-in Makefile rule such as .NOTPARALLEL it could
make the function detect 'all' rule as not the first rule which puts the
submitter in a position of having to override the warning which might
put them at a great risk of violating rule 17. As there are other
built-in Makefile rules it simply checks for the first character being a
'.'.

Also added missing not_reached() calls to noprompt_yes_or_no().