Try again fixing GitHub whining
It seems that it is NOT short-circuiting not being understood but some
other bug or mis-feature. It is safe to free a NULL pointer but in this
repo we always check for != NULL before freeing. However since we have
to check for line != NULL prior to dereferencing (though this is not
strictly necessary due to the fact we're also passing it to strchr()) we
cannot (because of GitHub) check for NULL before freeing it. That does
mean that if someone ended up setting it to NULL it would not be a valid
check any more. But since it is safe to free a NULL pointer it should
not matter which makes this check even sillier.