More work on mkiocccentry command line enhancement
Add to the macro MKIOCCCENTRY_DEV so that for now it at least will run the
tests on prog.c, Makefile and remarks.md and form the tarball even though
it's not complete (it won't ignore the right files, it doesn't do anything with
the other files etc. but this way at least the program won't fail when working
on the issue).

Changed MKIOCCCENTRY_VERSION to "1.2.4 2025-01-20".

Change mkiocccentry_test.sh to not test -d as it now sets -E which causes
a failure.