modify copyright to include both authors
Instead of removing the copyright and license, we
modify the copyright to include both authors.
Moreover as a fair mix, the order of the copyright authors
is the reverse order to the ".. tool was co-developed ..".
The quote at the bottom was moved to the top.

We propose this as a better model for jointly developed code.

The exit codes of `mkiocccentry.c` was also sequenced.

Performed `make release` to test the above under macOS.