Remove remarks.example.md
Instead this will be under next/ in the website (I already have done
some work on it but have not yet finished it). This is to make it like
the Makefile.example ans also so it can all be in one place (and more
easily linked to in the website - and if we wanted a html version it
would be easy to do too).