verify_submission_dir() now uses -R in ls
Function verify_submission_dir() now uses -R in ls and while there is a
line to read it checks if it matches the correct form. When the loop is done if
no correct line was found then it is an error (if it's < 0 from the beginning it
is also an error). However no files are copied over to the directory yet so this
cannot be fully tested.