collect_topdir_files() now copies files
collect_topdir_files() now copies additional files to the submission
directory. It requires obtaining absolute paths for the target file and the
source files because otherwise files might not be found (this does mean we have
to use getcwd() and have buffers of size PATH_MAX+1). There are numerous XXX
comments added pointing out that some of this has to be changed and perhaps
cleaned up (one such thing is that prog.c, Makefile and remarks.md will not be
checked when traversing the topdir but processed later on with the other files -
but there are other things that have to be done in addition to steps not yet
implemented). So although some of this will have to be changed some of it can
simply be moved and perhaps changed a bit (or adapted into what has to be done
depending on the steps). Depending on the file the permissions are different
(try.sh and try.alt.sh are the only ones allowed to be executable, for
example).